ettercap-ng: fix packet decoding (https://trac.macports.org/ticket/26526), disable universal (https://trac.macports.org/ticket/26060)
git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@75489 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ net/ettercap-ng/files/patch-src__protocols__ec_tcp.c
@@ -5,7 +5,7 @@
     
     opt_start = (u_char *)(tcp + 1);
 -   opt_end = (u_char *)((int)tcp + tcp->off * 4);
-+   opt_end = (u_char *)(tcp + tcp->off * 4);
++   opt_end = (u_char *)((unsigned long)tcp + tcp->off * 4);
  
     DECODED_LEN = (u_int32)(tcp->off * 4);
